Modified the prediction functions such that it's now possible to make predictions with a subset of covariance functions, which were used in "training" the model. For example, if the covariance function k is a sum covariance functions k_1 and k_2, of which the former represents a short-term correlation of a 1D time series, and the latter one a long-term correlation, we can now make predictions, which show only the short- or the long-term component of the overall prediction.
The interface for the component-wise predictions should be the
same with all the prediction functions (see the functions for details),
but there might still be some bugs lurking out there (especially with sparse
models), so consider yourself to be warned! 

Some demos of the prediction is to added in coming days.