Added the possibility to change the metric, which is used in calculating the distance between two input vectors in a covariance function.
The biggest change is that covariance function data structures now have
a field structure 'metric', which contains the relevant variables of the metric
and pointers to functions used by the metric, which are specified in their own
files (for now, only the euclidean metric is implemented) similarly as the
functions of covariance functions.

For now, this modification is done only for the squared exponential covariance,
but other traditional functions (exp, matern32 & matern52) are to be added
in near future. 

Also note that sexp is still backwards compatible with the old code meaning that
you don't have to use the metric structures, if you don't want to.

See demo_regression_metric for a short demo. It shows how euclidean metrics
are specified for different input components, and also how predictions
are made w.r.t different covariance functions.