Additional changes to "bug fix by O-P" on May 2016
Some additional changes, that were left out from Aki's commit "bug fix by O-P" on May 4 2016. All these changes are related to the same problem, which emerges when number of dimensions is larger than number of observations in test input vector 'xt': Because "length(xt)" always returns the length of the largest array dimension, it will return in this case the number of dimensions instead of the number of observations. Thus, "length(xt)" should be replaced by "size(xt,1)", if we want the number of observations. The same general problem may emerge also elsewhere...